Add Tuya lightingColorCtrl cluster and assign it to tuyaLight
