+++ src/converters/toZigbee.ts
@@ -2777,162 +2777,6 @@ export const easycode_auto_relock: Tz.Converter = {
         return {state: {auto_relock: value}};
     },
 };
-export const tuya_led_control: Tz.Converter = {
-    key: ["brightness", "color", "color_temp"],
-    options: [exposes.options.color_sync()],
-    convertSet: async (entity, key, value, meta) => {
-        if (
-            key === "brightness" &&
-            meta.state.color_mode === constants.colorModeLookup[2] &&
-            meta.message.color == null &&
-            meta.message.color_temp == null
-        ) {
-            const level = Number(value);
-
-            await entity.command(
-                "genLevelCtrl",
-                "moveToLevel",
-                {level, transtime: 0, optionsMask: 0, optionsOverride: 0},
-                utils.getOptions(meta.mapped, entity),
-            );
-
-            globalStore.putValue(entity, "brightness", level);
-
-            return {state: {brightness: level}};
-        }
-
-        if (key === "brightness" && utils.isNumber(meta.message.color_temp)) {
-            const level = Number(value);
-
-            await entity.command("lightingColorCtrl", "tuyaRgbMode", {enable: 0});
-            await entity.command(
-                "lightingColorCtrl",
-                "moveToColorTemp",
-                {
-                    colortemp: utils.mapNumberRange(meta.message.color_temp, 500, 154, 0, 254),
-                    transtime: 0,
-                    optionsMask: 0,
-                    optionsOverride: 0,
-                },
-                utils.getOptions(meta.mapped, entity),
-            );
-            await entity.command(
-                "genLevelCtrl",
-                "moveToLevel",
-                {level, transtime: 0, optionsMask: 0, optionsOverride: 0},
-                utils.getOptions(meta.mapped, entity),
-            );
-
-            globalStore.putValue(entity, "brightness", level);
-
-            const newState = {
-                brightness: level,
-                color_mode: constants.colorModeLookup[2],
-                color_temp: meta.message.color_temp,
-            };
-
-            return {state: libColor.syncColorState(newState, meta.state, entity, meta.options)};
-        }
-
-        if (key === "color_temp") {
-            utils.assertNumber(value, key);
-            const level = globalStore.getValue(entity, "brightness") || 100;
-
-            await entity.command("lightingColorCtrl", "tuyaRgbMode", {enable: 0});
-            await entity.command(
-                "lightingColorCtrl",
-                "moveToColorTemp",
-                {colortemp: utils.mapNumberRange(value, 500, 154, 0, 254), transtime: 0, optionsMask: 0, optionsOverride: 0},
-                utils.getOptions(meta.mapped, entity),
-            );
-            await entity.command(
-                "genLevelCtrl",
-                "moveToLevel",
-                {level, transtime: 0, optionsMask: 0, optionsOverride: 0},
-                utils.getOptions(meta.mapped, entity),
-            );
-
-            const newState = {
-                brightness: level,
-                color_mode: constants.colorModeLookup[2],
-                color_temp: value,
-            };
-
-            return {state: libColor.syncColorState(newState, meta.state, entity, meta.options)};
-        }
-
-        const zclData = {
-            brightness: globalStore.getValue(entity, "brightness") || 100,
-            // @ts-expect-error ignore
-            hue: utils.mapNumberRange(meta.state.color.h, 0, 360, 0, 254) || 100,
-            // @ts-expect-error ignore
-            saturation: utils.mapNumberRange(meta.state.color.s, 0, 100, 0, 254) || 100,
-            transtime: 0,
-        };
-
-        if (utils.isObject(value)) {
-            if (value.h) {
-                zclData.hue = utils.mapNumberRange(value.h, 0, 360, 0, 254);
-            }
-            if (value.hue) {
-                zclData.hue = utils.mapNumberRange(value.hue, 0, 360, 0, 254);
-            }
-            if (value.s) {
-                zclData.saturation = utils.mapNumberRange(value.s, 0, 100, 0, 254);
-            }
-            if (value.saturation) {
-                zclData.saturation = utils.mapNumberRange(value.saturation, 0, 100, 0, 254);
-            }
-            if (value.b) {
-                zclData.brightness = Number(value.b);
-            }
-            if (value.brightness) {
-                zclData.brightness = Number(value.brightness);
-            }
-            if (typeof value === "number") {
-                zclData.brightness = value;
-            }
-        }
-
-        if (meta.message.color != null) {
-            if (utils.isObject(meta.message.color)) {
-                if (meta.message.color.h) {
-                    zclData.hue = utils.mapNumberRange(meta.message.color.h, 0, 360, 0, 254);
-                }
-                if (meta.message.color.s) {
-                    zclData.saturation = utils.mapNumberRange(meta.message.color.s, 0, 100, 0, 254);
-                }
-                if (meta.message.color.b) {
-                    zclData.brightness = Number(meta.message.color.b);
-                }
-                if (meta.message.color.brightness) {
-                    zclData.brightness = Number(meta.message.color.brightness);
-                }
-            }
-        }
-
-        await entity.command("lightingColorCtrl", "tuyaRgbMode", {enable: 1});
-        await entity.command("lightingColorCtrl", "tuyaMoveToHueAndSaturationBrightness", zclData, utils.getOptions(meta.mapped, entity));
-
-        globalStore.putValue(entity, "brightness", zclData.brightness);
-
-        const newState = {
-            brightness: zclData.brightness,
-            color: {
-                h: utils.mapNumberRange(zclData.hue, 0, 254, 0, 360),
-                hue: utils.mapNumberRange(zclData.hue, 0, 254, 0, 360),
-                s: utils.mapNumberRange(zclData.saturation, 0, 254, 0, 100),
-                saturation: utils.mapNumberRange(zclData.saturation, 0, 254, 0, 100),
-            },
-            color_mode: constants.colorModeLookup[0],
-        };
-
-        return {state: libColor.syncColorState(newState, meta.state, entity, meta.options)};
-    },
-    convertGet: async (entity, key, meta) => {
-        await entity.read("lightingColorCtrl", ["currentHue", "currentSaturation", "tuyaBrightness", "tuyaRgbMode", "colorTemperature"]);
-    },
-};
 export const tuya_led_controller: Tz.Converter = {
     key: ["state", "color"],
     convertSet: async (entity, key, value, meta) => {
